move the component
